Generate request classes for all requests #20
Specifically, we were skipping request classes for AWS requests that had
no inputs, however, even with no inputs we still need a request class to
identifiy the request, HTTP method, etc.

So instead of skipping request classes, we simply add an overload that
allows the request to be invoked (on the relevent client class) without
specifying a request object (in which case we default-construct the
request instead).

Note, we use overloading here, instead of parameter defaulting, so that
the client class does not need to know the concrete request class (can
rely on pre-declarations only) which (should) allow clients to compile
quicker.  But either would be acceptable.